ksud: create hard link to ksu_susfs when kernel support susfs

+++ userspace/ksud/src/assets.rs
@@ -29,10 +29,11 @@ mod android {
         std::os::unix::fs::symlink("/data/adb/ksud", resetprop_link)?;
 
         // Create ksu_susfs -> ksud symlink (hard link)
-        let ksu_susfs = KSU_SUSFS;
-        let _ = std::fs::remove_file(ksu_susfs);
-        std::fs::hard_link("/data/adb/ksud", ksu_susfs)?;
-
+        if crate::android::susfs::show_version().is_ok() {
+            let ksu_susfs = KSU_SUSFS;
+            let _ = std::fs::remove_file(ksu_susfs);
+            std::fs::hard_link("/data/adb/ksud", ksu_susfs)?;
+        }
         Ok(())
     }
 }
